// src/service-api/generated/models/pickUserExcludeKeyofUserPassword.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * express-auth-postgresql
 * OpenAPI spec version: 1.0.0
 */
import type { PickUserExcludeKeyofUserPasswordRole } from './pickUserExcludeKeyofUserPasswordRole';

/**
 * From T, pick a set of properties whose keys are in the union K
 */
export interface PickUserExcludeKeyofUserPassword {
  id: number;
  email: string;
  name: string;
  role: PickUserExcludeKeyofUserPasswordRole;
  phone: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  county: string;
  postalCode: string;
  country: string;
  isEmailVerified: boolean;
  createdAt: string;
  updatedAt: string;
}